fix(aarch64): unblock storage selftest stall under tcg
- virtio: probe transports no longer drop-reset initialized devices
  (MmioTransport Drop writes STATUS=0, wiping the block device's vring
  before storage-selftest runs; mem::forget keeps probe-only transports
  resource-free instead)
- traps: BlockCurrentThreadOnReceive now honors deadline_ticks via
  block_current_on_receive_until, matching the x86_64 syscall paths
- selftest: cast DHCP xid field to u16 for to_be_bytes
- mmu: flush TLB (tlbi vmalle1) when switching page-table roots; a TTBR0
  write does not invalidate cached translations and all user spaces share
  ASID 0, so the second user task resumed with stale mappings from the
  first task's address space and executed the wrong image's code

+++ platform/aarch64/virt/src/selftest.rs
@@ -131,7 +131,7 @@ fn build_dhcp_discover(frame: &mut [u8; NET_BUFFER_BYTES], mac: &[u8; 6]) -> usi
     frame[14] = 0x45;
     frame[15] = 0x00;
     frame[16..18].copy_from_slice(&(ip_total as u16).to_be_bytes());
-    frame[18..20].copy_from_slice(&(DHCP_XID & 0xffff).to_be_bytes());
+    frame[18..20].copy_from_slice(&((DHCP_XID & 0xffff) as u16).to_be_bytes());
     frame[20] = 0x00;
     frame[21] = 0x00;
     frame[22] = 64;
